new dir structure for services types, update consent-api types, add empty dirs for pi, phi, keys das
