+++ packages/types/src/common/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -20,6 +20,7 @@
 export * from './IdAlphabet.js';
 export * from './Regex.js';
 export * from './SortOrder.js';
+export * from './Status.js';
 export * from './expand.js';
 export * from './keys.js';
 export * from './recursivePartial.js';
