+++ packages/types/src/entities/ConsentResearchParticipation.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -18,8 +18,6 @@
  */
 
 import { z } from 'zod';
-import { generateSchema } from '@anatine/zod-openapi';
-import type { SchemaObject } from 'openapi3-ts/oas31';
 
 import { ConsentQuestionId } from './ConsentQuestion.js';
 
@@ -30,19 +28,3 @@ export const ConsentResearchParticipationBase = z.object({
 	[RESEARCH_PARTICIPATION__CONTACT_INFORMATION]: z.boolean(),
 	[RESEARCH_PARTICIPATION__FUTURE_RESEARCH]: z.boolean(),
 });
-
-export const ConsentResearchParticipationRequest = ConsentResearchParticipationBase;
-export type ConsentResearchParticipationRequest = z.infer<
-	typeof ConsentResearchParticipationRequest
->;
-export const ConsentResearchParticipationRequestSchema: SchemaObject = generateSchema(
-	ConsentResearchParticipationRequest,
-);
-
-export const ConsentResearchParticipationResponse = ConsentResearchParticipationBase;
-export type ConsentResearchParticipationResponse = z.infer<
-	typeof ConsentResearchParticipationResponse
->;
-export const ConsentResearchParticipationResponseSchema: SchemaObject = generateSchema(
-	ConsentResearchParticipationResponse,
-);
