+++ packages/types/src/entities/InformedConsent.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -18,19 +18,9 @@
  */
 
 import { z } from 'zod';
-import { generateSchema } from '@anatine/zod-openapi';
-import type { SchemaObject } from 'openapi3-ts/oas31';
 
-import { ConsentQuestionId } from './ConsentQuestion.js';
+import { ConsentQuestionId } from './index.js';
 
-export const InformedConsentRequest = z.object({
+export const InformedConsentBase = z.object({
 	[ConsentQuestionId.enum.INFORMED_CONSENT__READ_AND_UNDERSTAND]: z.literal(true),
 });
-export type InformedConsentRequest = z.infer<typeof InformedConsentRequest>;
-export const InformedConsentRequestSchema: SchemaObject = generateSchema(InformedConsentRequest);
-
-export const InformedConsentResponse = z.object({
-	[ConsentQuestionId.enum.INFORMED_CONSENT__READ_AND_UNDERSTAND]: z.literal(true).optional(),
-});
-export type InformedConsentResponse = z.infer<typeof InformedConsentResponse>;
-export const InformedConsentResponseSchema: SchemaObject = generateSchema(InformedConsentResponse);
