+++ packages/types/src/entities/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -25,6 +25,7 @@ export * from './ConsentReleaseData.js';
 export * from './ConsentResearchParticipation.js';
 export * from './ConsentReviewSign.js';
 export * from './ConsentWizardProgress.js';
+export * from './InformedConsent.js';
 export * from './ParticipantIdentification.js';
 export * from './ParticipantResponse.js';
 export * from './User.js';
