+++ packages/types/src/services/consentApi/responses/ConsentReviewSign.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -17,5 +17,6 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-export * from './Status.js';
-export * from './IdAlphabet.js';
+import { ConsentReviewSignResponse as ConsentReviewSign } from '../../../entities/index.js';
+
+export const ConsentReviewSignResponse = ConsentReviewSign;
